requirements.txt
# Core Dependencies
# ==================

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# AI/ML APIs
google-generativeai>=0.8.0
openai>=1.0.0  # For OpenRouter compatibility

# Image/Video Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# HTTP & Networking
requests>=2.31.0
urllib3>=2.0.0
aiohttp>=3.9.0  # For async requests

# Environment & Configuration
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0  # Progress bars
colorama>=0.4.6  # Colored terminal output

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development Tools (optional)
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting
mypy>=1.5.0  # Type checking

# Logging & Monitoring
structlog>=23.1.0  # Structured logging

# Data Processing
pandas>=2.0.0  # Data manipulation
pydantic>=2.0.0  # Data validation